Remove use of reader when parsing
So far reader was used to deal with parsing (mutually) recursive
structures. This commit deals with the problem differently:

In addition to #inducedParser class method that defines how to parse
the structure it it introduces new (class) method `#asParser` following
the PPParser which is the to use in other `#inducedParser` methods to
parse constituents.

The key trick is that `#asParser` temporarily creates a delegate parser
(if not already) and then sets its contents to result if #inducedParser.
If this method is called recursively again (as part of #inducedParser),
then the delegate is already there and it is simply returned.

The advantage of this is that it is more readable and does not require
doing anything special if the structure may be recursive (which may not
be obvious).

While at it, make `JibProgramNode class >> #inducedParser` abstract and
get rid of `#adtSummands` - this just makes the code even simpler while
not significantly more work (only one or two new #inducedParser` were
added).