Remove use of `JibFnBody`
This commit refactors `Def_Fn` not to use `JibFnBody` and hold on
instructions directly. While at it, remove `body` instvar to `instrs`
(which is in line with upstream grammar/parser).

The class `JibFnBody` itself is kept around and will be deleted in a
commit with all other old stuff we decide to delete and build anew.