Refactor the use of "undefined value"
This commit refactors use of "undefined" value in the interpreter. It is
up to the type (`Ty` sub-instance) to answer new "undefined" value. Since
there's no single guaranteed "undefined" value, the checks in context's
`#get:` and `#set:to` had to be relaxed.

Note that this does not really matter because full-blown type unification
process, once in place, would also ensure the program is well-formed so
these checks at runtime won't be necessary.