Refactor `JobOp`
This commit refactors `JibOp` so it no longer an AST node but a subclass
of (new) `JibPrim`, a common base class for "ops" and `extern`s.

This seems to make more sense now with interpreter / type unification in
mind. A `JibPrim` knows its "signature" (a `Def_Val`) and knows how to
evaluate given argument values.