Rework parsing of Jib expression
This commit overhauls parsing of Jib expressions. Instead of using
"syntactic ADT" it reverts back to use standard way of defining parser
in PetitParser.

The main issue here is the left recursion in productions. This cannot be
handled by PetitParser and the only way is transform productions to
avoid that left recursion so one ends up in multiple productions for
the same thing depending on a context and all the beauty and simplicity
of "syntactic ADT" turns into a real mess.

It turned out to be much easier to implement this in separate (composite
parser) class where all rules are together. So this commit defines

    JibExp class >> asParser
        ^ JibParser new productionAt: #exp

and all the parsing is defined in `JibParser`. The rest of Jib AST
(definitions, instructions) still use "syntactic ADT" but I'm not sure
it is worth is.

With this change, a full `riscv64.ir` can be parsed.

Finally: it may be possible that PetitParser2 would be able to handle
recursion in this case as it tries hard to parse in LL(1) way and only
revert PEG-like parsing is necessary. Or at least that was the idea
we had with Jan decade ago when working on PetitCompiler. However, this
would require yet more experiments with PP2.