Add support for initializers in the interpreter
This commit add support for let's and register's initializers. These
initializers are executed before the main entry point.

This prompted some changes in the interpreter. Since there's no longer
a single "entry" function but rather a list of initializers followed
by the "entry" function, the interpreter maintains a worklist of
functions to execute (in `todo` instvar) and when the bottom-most context
returns, it simply executes next function in the worklist. Interpretation
finishes once all functions from worklist are executed and return (see
`JibInterpreter >> interpretEnd:`).

Another little complication comes from the fact that - for some unknown
reason - initializers do not have explicit `end` instruction. To deal
with this, after each instruction executed the interpreter checks if the
PC (of next instruction) is beyond the end of instruction sequence and
if it is, executes fabricated `end`. This way, the function return logic
is at single place (`JibInterpreter >> interpretEnd:`).