Allow to compare source locations
This commit adds `$=` and `#hash` to `JibSourceLoc` and subclasses,
allowing one to compare two locations.

Note, that that instances of `SourceLoc_Unknown` are equal only to
themselves (that is, if they're equal, then they must be identical).
This may be counter-intuitive, but will be helpful when implementing
implementing step-over and step-into on Sail source strata (see later
commit).