Add initial infrastructure for semantic analysis
This commits lays out a basic infrastructure for semantic analysis.
The basic idea is that `#doSemanticAnalysis` performs the analysis
somehow and creates "environment" for the program itself and for each
function. For now the environment does not have any functionality, this
will be added later as needed.

There may be multiple different (semantic) analyzers. This commit
introduces `JibProgramNaiveAnalyzer`. At the moment it is empty - the
actual analysis will be added in subsequent commits. Also, this "naive"
analyzer is expected to be superseded by more advanced one based on
Sprite and its bitvector length inference.