Use custom Z3 context when analyzing and interpreting Jib program
A Jib program may contain enums, structs and unions. These get
ultimately reduced to Z3 datatypes. For this reason, each program must
have its own Z3 context in which datatypes are defined. Also, this
context is the one to use when creating Z3 nodes representing "values".

This commit implements this by holding on a Z3 context from program's
environment and making sure semantic analysis and interpreter set this
context as current when analyzing and interpreting. This prompted

 1. creation of specialized `JibProgramEnvironment` that holds on such Z3
    context and
 2. basic "analysis" that fills in environment for program and
    defined `fn`s, `let`s and `register`s