Store "user-defined" datatypes in program environment
For the sake of interpretation and (later) formal reasoning Jib's user
types (enums, structs and unions) are represented as Z3 datatypes.

This commit adds `lookupSortForTypename:` & `#registerSort:forTypename:`
to environment objects. The former is to be used by interpreter to lookup
a datatype for given type name (id). The latter then by semantic analuzer
to associate Jib type with Z3 datatype.

Since Jib user types can be defined at program level only, mapping is
kept in a dictionary in `JibProgramEnvironment` and (function)
environments delegate to program's.