Add basic support for unit type in Sprite semantics and interpreter
This commit changes the way `%unit` sort is created. Instead of doing
manually by calling Z3 API, we add it to "Jib program Sprite prelude"
as `Data` and then rely on Sprite to create the datatype.

For some reason, using `%unit` as (Z3) datatype name did not work neither
did `()` as constructor name. Therefore I changed it to use `__unit__`
in both cases.