Rudiments of Extern
In this simple example, zand_vec goes directly to the bvand theory function.
This is, of course, wrong because Jib hardcodes its own set of ext names,
e.g. the name here should be 'and_bits'.

We can either add a map between Jib ext names and the theories (suitable
expanded if need be).  (Or we could even name the theory names from Jib
names, although I can see a number of disadvantages to the latter approach).